test(agent): move real-spawn hook test to smoke layer
The "SIGTERM-ignoring hook is SIGKILLed after the grace period" test
deliberately omits the injected spawner to exercise real signal behavior
via Bun.spawn. docs/test-readiness/test-system.md scopes *.unit.test.ts to
"no process spawning", so a real-spawn test belongs in the smoke layer —
matching the real-process preCheck precedent in cron-handler.smoke.test.ts.

Moved that one test verbatim (assertions unchanged) into a new
agent/src/hooks.smoke.test.ts. All other withCommandHooks tests stay in
hooks.unit.test.ts behind the injected spawner double.
